% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{bobeum}
\alias{bobeum}
\title{Genes of Blautia obeum}
\format{A list with 3210 elements}
\source{
\url{https://img.jgi.doe.gov/cgi-bin/m/main.cgi?section=TaxonDetail&page=taxonDetail&taxon_oid=650377967}
}
\usage{
bobeum
}
\description{
Fasta file containing the genes of B. obeum A2-162, retrieved from IMG
This file contains only the coding regions of B. obeum, with one sequence
per gene.
}
\keyword{datasets}
