% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{efaecalis}
\alias{efaecalis}
\title{Genes of Enterococcus faecalis}
\format{A list with 3094 elements}
\source{
\url{https://img.jgi.doe.gov/cgi-bin/m/main.cgi?section=TaxonDetail&page=taxonDetail&taxon_oid=650377929}
}
\usage{
efaecalis
}
\description{
Fasta file containing the genes of E. faecalis 62, retrieved from IMG
This file contains only the coding regions of E. faecalis, with one sequence
per gene.
}
\keyword{datasets}
