% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fprausnitzii}
\alias{fprausnitzii}
\title{Genes of Faecalibacterium prausnitzii}
\format{A list with 3553 elements}
\source{
\url{https://img.jgi.doe.gov/cgi-bin/m/main.cgi?section=TaxonDetail&page=taxonDetail&taxon_oid=645951831}
}
\usage{
fprausnitzii
}
\description{
Fasta file containing the genes of F. prausnitzii A2-165, retrieved from IMG
This file contains only the coding regions of F. prausnitzii, with one sequence
per gene.
}
\keyword{datasets}
