% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ljohnsonii}
\alias{ljohnsonii}
\title{Genes of Lactobacillus johnsonii}
\format{A list with 1941 elements}
\source{
\url{https://img.jgi.doe.gov/cgi-bin/m/main.cgi?section=TaxonDetail&page=taxonDetail&taxon_oid=637000140}
}
\usage{
ljohnsonii
}
\description{
Fasta file containing the genes of L. johnsonii NCC 533, retrieved from IMG
This file contains only the coding regions of L. johnsonii, with one sequence
per gene.
}
\keyword{datasets}
