% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadPasilla}
\alias{loadPasilla}
\title{Define the composition of all samples in a metatranscriptomics dataset}
\usage{
loadPasilla()
}
\description{
\code{loadPasilla} imports the Pasilla dataset from the corresponding package.
This is an internal function and should not be called directly by the user, only internally by the \code{simulateSingleGenome} function
}
