% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocs.R
\docType{data}
\name{rintestinalis}
\alias{rintestinalis}
\title{Genes of Roseburia intestinalis}
\format{A list with 4865 elements}
\source{
\url{https://img.jgi.doe.gov/cgi-bin/m/main.cgi?section=TaxonDetail&page=taxonDetail&taxon_oid=2562617159}
}
\usage{
rintestinalis
}
\description{
Fasta file containing the genes of R. intestinalis L1-82, retrieved from IMG
This file contains only the coding regions of R. intestinalis, with one sequence
per gene.
}
\keyword{datasets}
