% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sampleReads}
\alias{sampleReads}
\title{Scale the number of simulated reads to the actual number per genome}
\usage{
sampleReads(countMatrix, readNumber, seed = 42)
}
\arguments{
\item{countMatrix}{Gene count matrix to adjust for the number of total reads per sample.}

\item{readNumber}{A number, indicating the total samples (cases and controls).}

\item{seed}{Random seed for the sampling.}
}
\value{
A gene count matrix, in which the sum of the reads per sample is equal to the
    number of reads specified for a specific genome
}
\description{
\code{sampleReads} returns a composition matrix scaled to the actual reads
   assigned to a genome. The function \code{simulateSingleGenome} overestimates the
   number of reads, because of this we sample with replacement to obtain the number
   of reads per genome required.
   This is an internal function and should not be called directly by the user, only
   internally by the \code{simulateSingleGenome} function
}
\seealso{
\code{\link{simulateSingleGenome}}
}
