% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_distribution_genomes.R
\name{compositionGenomesMetaT}
\alias{compositionGenomesMetaT}
\title{Define the species composition of all samples in a metatranscriptomics dataset}
\usage{
compositionGenomesMetaT(composition = c("custom", "empirical", "even"),
  empiricalSeed = NULL, genomes, compositionMatrix = NULL, nReads = 1e+06,
  nReplicates = 10, seed = 42)
}
\arguments{
\item{composition}{A character string indicating which method to use to determine
the composition of the matrix. Any of "custom", "empirical" or "even" can be used.
See "details"}

\item{empiricalSeed}{A single number or a numeric vector of length equal to
nReplicates. Indicates the random seed to assign the reads to different species
in each sample. If NULL or a single number, the same seed will be applied to all
samples so that they will have the same composition, with differences only due to
random sampling}

\item{genomes}{Character vector of genome names or genome IDs for the genomes to
include in the simulation}

\item{compositionMatrix}{A composition matrix predefined by the user, in which rows
represent species and columns represent samples. Id user provides a composition
matrix, this function only performs the random sampling to scale to the number of
reads desired. Defaults to NULL}

\item{nReads}{An integer, indicating the number of reads to simulate per sample.}

\item{nReplicates}{An integer, indicating the total samples (cases and controls).}

\item{seed}{An integer, sets the random seed for the read distribution.}
}
\value{
A microbial composition matrix of \code{nReplicates} columns and
    \code{nrow(genomes)} rows.
}
\description{
\code{compositionGenomesMetaT} returns a composition matrix with rows as species/
    genomes and columns as samples (cases or controls)
}
\examples{
# define a list of genomes to simulate
genomesToSimulate <- c("F. prausnitzii", "R. intestialis", "L. lactis", "E. faecalis",
                       "R. obeum")
# obtain the empirical composition matrix for this 5 species
compositionGenomesMetaT(composition="empirical", empiricalSeed=1,
                        genomes=genomesToSimulate, nReads=500000,nReplicates=10)
}
