% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{empiricalComposition}
\alias{empiricalComposition}
\title{Define the composition of all samples in a metatranscriptomics dataset}
\usage{
empiricalComposition(empiricalSeed = NULL, genomes, nReplicates)
}
\arguments{
\item{empiricalSeed}{A single number or a numeric vector of length equal to nReplicates. Indicates the random seed to assign the reads to different species in each sample. If NULL or a single number, the same seed will be applied to all samples so that they will have the same composition, with differences only due to random sampling.}

\item{genomes}{Character vector of genome names or genome IDs for the genomes to include in the simulation.}

\item{nReplicates}{A number, indicating the total samples (cases and controls).}
}
\value{
A microbial composition matrix of \code{nReplicates} columns and \code{nrow(genomes)} rows.
}
\description{
\code{empiricalComposition} returns a composition matrix with rows as species/genomes and columns as samples (cases or controls). This composition matrix follows an empirical distribution which has been fitted from real metatranscriptomics datasets (Metatrans paper).
This is an internal function and should not be called directly by the user, only internally by the \code{compositionGenomesMetaT} function
}
\examples{
genomesToSimulate <- c("F. prausnitzii", "R. intestialis", "L. lactis", "E. faecalis", "R. obeum")
empiricalComposition(empiricalSeed=1, genomes=genomesToSimulate, nReplicates=10)
}
\seealso{
compositionGenomesMetaT
}
